Refactors player properties with getter/setter macro
Replaces individual getter and setter methods for player properties (health, speed, size, crit chance, crit multiplier) with a unified GETTER_SETTER macro.

This improves code readability and reduces boilerplate code.
Also, the assert include path was changed.
