+++ src/imgui/fpscounter.hpp
@@ -1,7 +1,7 @@
 #pragma once
 #include "menu.hpp"
 #include <imgui.h>
-#include "../helper/assert.hpp"
+#include "../utils/assert.hpp"
 
 constexpr unsigned int NOEDIT_FLAGS = ImGuiWindowFlags_NoCollapse
 | ImGuiWindowFlags_NoSavedSettings
