+++ src/utils/getter_setter.hpp
@@ -1,0 +1,43 @@
+#pragma once
+
+
+/**
+ * @brief A utility to build a get method on top of a property.
+ *
+ * @param type The type of the property.
+ * @param name The machine-readable name of the property.
+ */
+#ifndef GETTER_IMPL
+#define GETTER_IMPL(type, method, name) \
+    /** Retrieves the name value for the object. @return The name value of the object. */ \
+    type Get##method() const { \
+        return name; \
+    }
+#endif
+
+/**
+ * @brief A utility to build a set method on top of a property.
+ * 
+ * @param type The type of the property.
+ * @param name The machine-readable name of the property.
+ */
+#ifndef SETTER_IMPL
+#define SETTER_IMPL(type, method, name) \
+    /** Sets the name value for the object. @param value The value of which to set name to. */ \
+    void Set##method(type value) { \
+        name = value; \
+    }
+#endif
+
+
+#ifndef GETTER_SETTER
+/**
+ * @brief A utility to build get and set methods on top of a property.
+ *
+ * @param type The type of the property.
+ * @param name The machine-readable name of the property.
+ */
+#define GETTER_SETTER(type, method, name) \
+    GETTER_IMPL(type, method, name) \
+    SETTER_IMPL(type, method, name)
+#endif
